+++ changelog.md
@@ -1,3 +1,11 @@
+# 1.1.1
+- Fixed a bug where all of the quotes could appear at once.
+- Fixed my lazy code to actually make the "Rare Quote" rare.
+- Change chance for rare quote from 0.1% to 0.01%
+- Added settings to Enable or Disable the button in the Main Menu.
+- Set dependency to now require Node-IDs v1.15.0 and up.
+- Added a few new quotes.
+
 # 1.1.0
 - Added Android and Mac support.
 - Fixed button overlapping in pause menu.